improvement

+++ PurchaseShared/Models/VendorStat.cs
@@ -6,7 +6,7 @@ namespace PurchaseShared.Models
         public VendorStat()
         {
         }
-        public VendorStat(string id, string name, double totalAmounts, string type)
+        public VendorStat(string id, string name, decimal totalAmounts, string type)
         {
             Id = id; Name = name; TotalAmounts = totalAmounts;
             Type = type;
@@ -14,7 +14,7 @@ namespace PurchaseShared.Models
         public string Id { get; set; }
         public string Name { get; set; }
         public string Type { get; set; }
-        public double TotalAmounts { get; set; }
+        public decimal TotalAmounts { get; set; }
 
         public string ShortName => Name.Length > 5 ? Name.Substring(0,5) : Name;
